requirements.txt
amqp==5.3.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
filelock==3.16.1
fsspec==2024.12.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
